requirements.txt
autopep8==1.5.7
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
et-xmlfile==1.1.0
idna==3.3
inflect==5.3.0
joblib==1.1.0
ndjson==0.3.1
nltk==3.6.5
openpyxl==3.0.9
psycopg2-binary==2.9.1
pycodestyle==2.8.0
python-dotenv==0.19.1
regex==2021.10.21
requests==2.26.0
toml==0.10.2
tqdm==4.62.3
urllib3==1.26.7
